PC登录较权完成

+++ health_web/health_pc/src/main/java/com/itheima/health/controller/SetmealController.java
@@ -45,10 +45,13 @@ public class SetmealController {
     })
     @Swagger2CommonConfiguration
     public Result add(@RequestBody SetmealDTO setmealDTO) {
-        RedisUtil.removeSetMember(RedisConstant.ALL_SETMEAL_PIC_SET, setmealDTO.getImg());
-        if (!RedisUtil.existsKey(RedisConstant.SINGLE_PIC + setmealDTO.getImg())) {
-            AliyunUtils.deleteFile(setmealDTO.getImg());
-            return new Result(false, "图片上传超时！请重新上传");
+        if (!setmealDTO.getImg().equals(setmealService.getById(setmealDTO.getId()).getImg())) {
+            RedisUtil.removeSetMember(RedisConstant.ALL_SETMEAL_PIC_SET, setmealDTO.getImg());
+            if (!RedisUtil.existsKey(RedisConstant.SINGLE_PIC + setmealDTO.getImg())) {
+                AliyunUtils.deleteFile(setmealDTO.getImg());
+                return new Result(false, "图片上传超时！请重新上传");
+            }
+            AliyunUtils.deleteFile(setmealService.getById(setmealDTO.getId()).getImg());
         }
         setmealService.saveUpdate(setmealDTO);
         RedisUtil.delete(RedisConstant.SINGLE_PIC + setmealDTO.getImg());
